+++ docs/changelog.rst
@@ -1,6 +1,16 @@
 Changelog
 =========
 
+0.1.3 (2023-07-03)
+------------------
+
+Changed
+~~~~~~~
+
+-  :meth:`yapw.clients.Async.connection_open_error_callback`: Don't retry if the error is an authentication (e.g. incorrect username or password) or access denied (e.g. non-existent virtual host) error.
+-  :meth:`yapw.clients.Async.connection_close_callback`: Change log level from ``ERROR`` to ``WARNING``.
+-  :meth:`yapw.clients.Async` log messages format errors with ``%r`` instead of ``%s``, to include information like the exception class.
+
 0.1.2 (2023-07-02)
 ------------------